requirements.txt
fastapi==0.103.1
psycopg2-binary==2.9.5
SQLAlchemy==1.4.44
uvicorn==0.19.0
email-validator==1.3.0
emails==0.6
celery
celery[redis]
alembic==1.8.1
bcrypt==4.0.1
redis
pydantic==2.3.0
pydantic-settings
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
starlette
passlib
smtplib